Setup a baseline Codespaces configuration
This uses the universal image, which whilst very large is either already
on the hosts the Codespace runs on, or doesn't take long to clone. It
then sets up a Postgres database, with a custom user for the Codespace
account. After the environment is up, it runs `bin/setup` (which itself
has some additions to help the environment work).

The configuration diverges fairly significantly from other examples,
none of which worked in practice.